requirements.txt
beautifulsoup4==4.12.3
certifi==2025.8.3
charset-normalizer==3.3.2
idna==3.7
python-dotenv==1.1.0
selenium==4.34.2
requests==2.32.3
soupsieve==2.5
typing_extensions==4.14.1
urllib3==2.5.0